Implemented basic database architecture
I need to migrate away from the unwieldy json files if I'm going to try to keep track of timestamps and more user data in the future, so I'm moving the data storage system over to a SQL database. As it stands, this is the basic architecture, although it still uses the JSON files.
I also rearranged the filesystem to be a little more organized, and implemented some library support code (a custom decorator class) and the start of a data layer (which utilizes this decorator class)
Finally, the class inside the ModToolsCog file is once again changing its name, and i tried to methodize the data layer out a bit more to make the transition from JSON to SQL a little smoother in the future.