db architecture and warn process transfer
Completed the new database architecture, swapping away from the json architecture completely. In the process, also implemented a way to completely separate the data layer from the business layer, so as to have minimal coupling.

Additionally, completed the original enhancement goal of decaying warnings. When a warn action is called, before processing the action, it checks the given user for warnings and removes outdated ones (assuming the warning duration configruation, `warning_duration_days`, is not set to 0).